+++ radiant_mlhub/__init__.py
@@ -2,8 +2,9 @@ __all__ = [
     "__version__",
     "Collection",
     "Dataset",
+    "MLModel",
     "get_session"
 ]
 from .__version__ import __version__
-from .models import Collection, Dataset
+from .models import Collection, Dataset, MLModel
 from .session import get_session