+++ radiant_mlhub/models/__init__.py
@@ -1,0 +1,10 @@
+"""Extensions of the `PySTAC <https://pystac.readthedocs.io/en/latest/>`_ classes that provide convenience methods for interacting
+with the `Radiant MLHub API <https://docs.mlhub.earth/#radiant-mlhub-api>`_."""
+__all__ = [
+    "Collection",
+    "Dataset",
+    "MLModel",
+]
+from .collection import Collection
+from .dataset import Dataset
+from .ml_model import MLModel