+++ test/test_models.py
@@ -1,17 +1,21 @@
 import json
-import pytest
 import re
-import pystac.item
 from pathlib import Path
-from urllib.parse import urljoin, urlsplit, parse_qs
-from typing import Any, cast, Dict, Iterator, TYPE_CHECKING
+from typing import TYPE_CHECKING, Any, Dict, Iterator, cast
+from urllib.parse import parse_qs, urljoin, urlsplit
 
-from radiant_mlhub.models import Collection, Dataset
+import pystac.item
+import pytest
+from pystac.asset import Asset
+from pystac.link import Link
+from radiant_mlhub.models import Collection, Dataset, MLModel
 
 from . import util
+
 if TYPE_CHECKING:
-    from requests_mock import Mocker as Mocker_Type
     from pathlib import Path as Path_Type
+
+    from requests_mock import Mocker as Mocker_Type
 
 
 class TestCollection:
@@ -405,3 +409,43 @@ class TestDatasetNoProfile:
         history = requests_mock.request_history
         assert len(history) == 2
         assert f"key={api_key}" in history[1].url
+
+
+class TestMLModel:
+
+    @pytest.mark.vcr
+    def test_list_ml_models(self) -> None:
+        """MLModel.list returns a list of MLModel instances."""
+        ml_models = MLModel.list()
+        assert isinstance(ml_models[0], MLModel)
+
+    @pytest.mark.vcr
+    def test_fetch_ml_model_by_id(self) -> None:
+        expect_id = 'model-cyclone-wind-estimation-torchgeo-v1'
+        ml_model = MLModel.fetch(expect_id)
+        assert isinstance(ml_model, MLModel)
+        assert ml_model.id == expect_id
+        assert len(ml_model.links) > 0
+        assert len(ml_model.bbox) > 0
+        assert ml_model.collection_id == 'model-cyclone-wind-estimation-torchgeo'
+        assert ml_model.properties.get('title') == 'Tropical Cyclone Wind Estimation Model'
+        assert ml_model.properties.get('sci:doi') == '10.5281/zenodo.5773331'
+        assert ml_model.properties.get('ml-model:type') == 'ml-model'
+        assert len(ml_model.properties.get('providers'))
+        assert len(ml_model.properties.get('sci:publications'))
+        assert isinstance(ml_model.assets.get('inferencing-compose'), Asset)
+        assert isinstance(ml_model.assets.get('inferencing-checkpoint'), Asset)
+        assert len(ml_model.links) > 0
+        assert isinstance(ml_model.links[0], Link)
+
+    @pytest.mark.vcr
+    def test_ml_model_resolves_links(self) -> None:
+        expect_id = 'model-cyclone-wind-estimation-torchgeo-v1'
+        ml_model = MLModel.fetch(expect_id)
+        for link in ml_model.links:
+            assert isinstance(link, Link)
+            assert link.href is not None
+
+    @pytest.mark.skip(reason="MLModel get by doi is not implemented")
+    def test_get_ml_model_by_doi(self, requests_mock: "Mocker_Type", root_url: str) -> None:
+        pass